split basemodel into several files, plotting goes in analysis module
* basestring is dropped, only using str (as far as I can see a relic
  from py2/py3 compat)
* matplotlib determines itself the backend. matplotlib and cartopy
  always required.
